+++ wsi_service/api/v1/slides.py
@@ -9,6 +9,7 @@ from wsi_service.custom_models.queries import (
     ImageFormatsQuery,
     ImagePaddingColorQuery,
     ImageQualityQuery,
+    PluginQuery,
     ZStackQuery,
 )
 from wsi_service.custom_models.responses import ImageRegionResponse, ImageResponses
@@ -22,15 +23,21 @@ from wsi_service.utils.app_utils import (
     validate_image_size,
     validate_image_z,
 )
+from wsi_service.utils.image_utils import (
+    check_complete_region_overlap,
+    check_complete_tile_overlap,
+    get_extended_region,
+    get_extended_tile,
+)
 
 
 def add_routes_slides(app, settings, slide_manager):
     @app.get("/slides/{slide_id}/info", response_model=SlideInfo, tags=["Main Routes"])
-    async def _(slide_id: str):
+    async def _(slide_id: str, plugin: str = PluginQuery):
         """
         Get metadata information for a slide given its ID
         """
-        return await slide_manager.get_slide_info(slide_id, slide_info_model=SlideInfo)
+        return await slide_manager.get_slide_info(slide_id, slide_info_model=SlideInfo, plugin=plugin)
 
     @app.get(
         "/slides/{slide_id}/thumbnail/max_size/{max_x}/{max_y}",
@@ -48,6 +55,7 @@ def add_routes_slides(app, settings, slide_manager):
         ),
         image_format: str = ImageFormatsQuery,
         image_quality: int = ImageQualityQuery,
+        plugin: str = PluginQuery,
     ):
         """
         Get slide thumbnail image  given its ID.
@@ -59,7 +67,7 @@ def add_routes_slides(app, settings, slide_manager):
         When tiff is specified as output format the raw data of the image is returned.
         """
         validate_image_request(image_format, image_quality)
-        slide = await slide_manager.get_slide(slide_id)
+        slide = await slide_manager.get_slide(slide_id, plugin=plugin)
         thumbnail = await slide.get_thumbnail(max_x, max_y)
         return make_response(slide, thumbnail, image_format, image_quality)
 
@@ -75,6 +83,7 @@ def add_routes_slides(app, settings, slide_manager):
         max_y: int = Path(None, example=100, description="Maximum height of label image"),
         image_format: str = ImageFormatsQuery,
         image_quality: int = ImageQualityQuery,
+        plugin: str = PluginQuery,
     ):
         """
         Get the label image of a slide given its ID.
@@ -86,7 +95,7 @@ def add_routes_slides(app, settings, slide_manager):
         When tiff is specified as output format the raw data of the image is returned.
         """
         validate_image_request(image_format, image_quality)
-        slide = await slide_manager.get_slide(slide_id)
+        slide = await slide_manager.get_slide(slide_id, plugin=plugin)
         label = await slide.get_label()
         label.thumbnail((max_x, max_y), Image.ANTIALIAS)
         return make_response(slide, label, image_format, image_quality)
@@ -103,6 +112,7 @@ def add_routes_slides(app, settings, slide_manager):
         max_y: int = Path(None, example=100, description="Maximum height of macro image"),
         image_format: str = ImageFormatsQuery,
         image_quality: int = ImageQualityQuery,
+        plugin: str = PluginQuery,
     ):
         """
         Get the macro image of a slide given its ID.
@@ -114,7 +124,7 @@ def add_routes_slides(app, settings, slide_manager):
         When tiff is specified as output format the raw data of the image is returned.
         """
         validate_image_request(image_format, image_quality)
-        slide = await slide_manager.get_slide(slide_id)
+        slide = await slide_manager.get_slide(slide_id, plugin=plugin)
         macro = await slide.get_macro()
         macro.thumbnail((max_x, max_y), Image.ANTIALIAS)
         return make_response(slide, macro, image_format, image_quality)
@@ -134,8 +144,10 @@ def add_routes_slides(app, settings, slide_manager):
         size_y: int = Path(None, gt=0, example=1024, description="Height of requested region"),
         image_channels: List[int] = ImageChannelQuery,
         z: int = ZStackQuery,
+        padding_color: str = ImagePaddingColorQuery,
         image_format: str = ImageFormatsQuery,
         image_quality: int = ImageQualityQuery,
+        plugin: str = PluginQuery,
     ):
         """
         Get a region of a slide given its ID and by providing the following parameters:
@@ -164,6 +176,10 @@ def add_routes_slides(app, settings, slide_manager):
         * `z` - The region endpoint also offers the selection of a layer in a Z-Stack by setting the index z.
         Default is z=0.
 
+        * `padding_color` - Background color as 24bit-hex-string with leading #,
+        that is used when image region contains whitespace when out of image extent. Default is white.
+        Only works for 8-bit RGB slides, otherwise the background color is black.
+
         * `image_format` - The image format can be selected. Formats include jpeg, png, tiff, bmp, gif.
         When tiff is specified as output format the raw data of the image is returned.
         Multi-channel images can also be represented as RGB-images (mostly for displaying reasons in the viewer).
@@ -173,14 +189,20 @@ def add_routes_slides(app, settings, slide_manager):
         * `image_quality` - The image quality can be set for specific formats,
         e.g. for the jpeg format a value between 0 and 100 can be selected. Default is 90.
         """
+        vp_color = validate_hex_color_string(padding_color)
         validate_image_request(image_format, image_quality)
         validate_image_size(size_x, size_y)
-        slide = await slide_manager.get_slide(slide_id)
+        slide = await slide_manager.get_slide(slide_id, plugin=plugin)
         slide_info = await slide.get_info()
         validate_image_level(slide_info, level)
         validate_image_z(slide_info, z)
         validate_image_channels(slide_info, image_channels)
-        image_region = await slide.get_region(level, start_x, start_y, size_x, size_y, padding_color=None)
+        if check_complete_region_overlap(slide_info, level, start_x, start_y, size_x, size_y):
+            image_region = await slide.get_region(level, start_x, start_y, size_x, size_y, padding_color=vp_color, z=z)
+        else:
+            image_region = await get_extended_region(
+                slide.get_region, slide_info, level, start_x, start_y, size_x, size_y, padding_color=vp_color, z=z
+            )
         return make_response(slide, image_region, image_format, image_quality, image_channels)
 
     @app.get(
@@ -199,6 +221,7 @@ def add_routes_slides(app, settings, slide_manager):
         padding_color: str = ImagePaddingColorQuery,
         image_format: str = ImageFormatsQuery,
         image_quality: int = ImageQualityQuery,
+        plugin: str = PluginQuery,
     ):
         """
         Get a tile of a slide given its ID and by providing the following parameters:
@@ -226,6 +249,7 @@ def add_routes_slides(app, settings, slide_manager):
 
         * `padding_color` - Background color as 24bit-hex-string with leading #,
         that is used when image tile contains whitespace when out of image extent. Default is white.
+        Only works for 8-bit RGB slides, otherwise the background color is black.
 
         * `image_format` - The image format can be selected. Formats include jpeg, png, tiff, bmp, gif.
         When tiff is specified as output format the raw data of the image is returned.
@@ -238,10 +262,15 @@ def add_routes_slides(app, settings, slide_manager):
         """
         vp_color = validate_hex_color_string(padding_color)
         validate_image_request(image_format, image_quality)
-        slide = await slide_manager.get_slide(slide_id)
+        slide = await slide_manager.get_slide(slide_id, plugin=plugin)
         slide_info = await slide.get_info()
         validate_image_level(slide_info, level)
         validate_image_z(slide_info, z)
         validate_image_channels(slide_info, image_channels)
-        image_tile = await slide.get_tile(level, tile_x, tile_y, padding_color=vp_color, z=z)
+        if check_complete_tile_overlap(slide_info, level, tile_x, tile_y):
+            image_tile = await slide.get_tile(level, tile_x, tile_y, padding_color=vp_color, z=z)
+        else:
+            image_tile = await get_extended_tile(
+                slide.get_tile, slide_info, level, tile_x, tile_y, padding_color=vp_color, z=z
+            )
         return make_response(slide, image_tile, image_format, image_quality, image_channels)