+++ wsi_service/custom_models/queries.py
@@ -20,6 +20,7 @@ ImagePaddingColorQuery = Query(
     example="#FFFFFF",
     description="""Background color as 24bit-hex-string with leading #,
     that is used when image tile contains whitespace when out of image extent. Default is white.
+    Only works for 8-bit RGB slides, otherwise the background color is black.
     """,
 )